+++ lib/main_db/new/providers/index.dart
@@ -1,4 +1,5 @@
 library;
 
 export 'db_history_provider.dart';
+export 'close_sync_tracking_provider.dart';
 export 'main_store_manager_provider.dart';