+++ lib/main_db/old/provider/main_store_close_sync_controller.dart
@@ -33,6 +33,7 @@ class MainStoreCloseSyncController {
   bool _pendingSnapshotUploadPromptOnClose = false;
   Completer<bool>? _closeStoreUploadDecision;
 
+  // 
   Future<void> tryUploadSnapshotBeforeClose({
     required MainStoreManager manager,
     required String logTag,